simplify mpu module part3
