+++ NLP/GLM/loss.txt
@@ -1,1 +1,1 @@
-7.982724189758301
+8.1207914352417
