+++ NLP/GLM/mpu/cross_entropy.py
@@ -16,9 +16,9 @@
 
 import oneflow as flow
 
-from .initialize import get_model_parallel_group
-from .initialize import get_model_parallel_rank
-from .initialize import get_model_parallel_world_size
+from .distribute import get_model_parallel_group
+from .distribute import get_model_parallel_rank
+from .distribute import get_model_parallel_world_size
 from .utils import VocabUtility
