+++ NLP/GLM/mpu/data.py
@@ -15,10 +15,6 @@
 
 import oneflow as flow
 
-from .initialize import get_model_parallel_group
-from .initialize import get_model_parallel_rank
-from .initialize import get_model_parallel_src_rank
-
 
 _MAX_DATA_DIM = 5
