+++ NLP/GLM/mpu/distribute.py
@@ -1,0 +1,135 @@
+# coding=utf-8
+# Copyright (c) 2019, NVIDIA CORPORATION.  All rights reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+
+"""Model and data parallel groups."""
+
+import oneflow as flow
+
+from .utils import ensure_divisibility
+
+
+# Model parallel group that the current rank belongs to.
+_MODEL_PARALLEL_GROUP = None
+# Data parallel group that the current rank belongs to.
+_DATA_PARALLEL_GROUP = None
+
+
+# def initialize_model_parallel(model_parallel_size_):
+#     """
+#     Initialize model data parallel groups.
+
+#     Arguments:
+#         model_parallel_size: number of GPUs used to parallelize model.
+
+#     Let's say we have a total of 8 GPUs denoted by g0 ... g7 and we
+#     use 2 GPUs to parallelize the model. The present function will
+#     create 4 model parallel groups and 2 data parallel grous as:
+#         4 model parallel groups:
+#             [g0, g1], [g2, g3], [g4, g5], [g6, g7]
+#         2 data parallel groups:
+#             [g0, g2, g4, g6], [g1, g3, g5, g7]
+#     Note that for efficiency, the caller should make sure adjacent ranks
+#     are on the same DGX box. For example if we are using 2 DGX-1 boxes
+#     with a total of 16 GPUs, rank 0 to 7 belong to the first box and
+#     ranks 8 to 15 belong to the second box.
+#     """
+#     if flow.distributed.get_rank() == 0:
+#         print('> initializing model parallel with size {}'.format(
+#             model_parallel_size_))
+#     # Get world size and rank. Ensure some consistencies.
+#     assert flow.distributed.is_initialized()
+#     world_size = flow.distributed.get_world_size()
+#     model_parallel_size = min(model_parallel_size_, world_size)
+#     ensure_divisibility(world_size, model_parallel_size)
+#     rank = flow.distributed.get_rank()
+
+#     # Build the data parallel groups.
+#     global _DATA_PARALLEL_GROUP
+#     assert _DATA_PARALLEL_GROUP is None, \
+#         'data parallel group is already initialized'
+#     for i in range(model_parallel_size):
+#         ranks = range(i, world_size, model_parallel_size)
+#         group = flow.distributed.new_group(ranks)
+#         if i == (rank % model_parallel_size):
+#             _DATA_PARALLEL_GROUP = group
+
+#     # Build the model parallel groups.
+#     global _MODEL_PARALLEL_GROUP
+#     assert _MODEL_PARALLEL_GROUP is None, \
+#         'model parallel group is already initialized'
+#     for i in range(world_size // model_parallel_size):
+#         ranks = range(i * model_parallel_size,
+#                       (i + 1) * model_parallel_size)
+#         group = flow.distributed.new_group(ranks)
+#         if i == (rank // model_parallel_size):
+#             _MODEL_PARALLEL_GROUP = group
+
+
+def model_parallel_is_initialized():
+    """Check if model and data parallel groups are initialized."""
+    if _MODEL_PARALLEL_GROUP is None or _DATA_PARALLEL_GROUP is None:
+        return False
+    return True
+
+
+def get_model_parallel_group():
+    """Get the model parallel group the caller rank belongs to."""
+    assert _MODEL_PARALLEL_GROUP is not None, \
+        'model parallel group is not initialized'
+    return _MODEL_PARALLEL_GROUP
+
+
+def get_data_parallel_group():
+    """Get the data parallel group the caller rank belongs to."""
+    assert _DATA_PARALLEL_GROUP is not None, \
+        'data parallel group is not initialized'
+    return _DATA_PARALLEL_GROUP
+
+
+def get_model_parallel_world_size():
+    """Return world size for the model parallel group."""
+    return flow.distributed.get_world_size(group=get_model_parallel_group())
+
+
+def get_model_parallel_rank():
+    """Return my rank for the model parallel group."""
+    return flow.distributed.get_rank(group=get_model_parallel_group())
+
+
+def get_model_parallel_src_rank():
+    """Calculate the global rank corresponding to a local rank zeor
+    in the model parallel group."""
+    global_rank = flow.distributed.get_rank()
+    local_world_size = get_model_parallel_world_size()
+    return (global_rank // local_world_size) * local_world_size
+
+
+def get_data_parallel_world_size():
+    """Return world size for the data parallel group."""
+    return flow.distributed.get_world_size(group=get_data_parallel_group())
+
+
+def get_data_parallel_rank():
+    """Return my rank for the data parallel group."""
+    return flow.distributed.get_rank(group=get_data_parallel_group())
+
+
+def destroy_model_parallel():
+    """Set the groups to none."""
+    global _MODEL_PARALLEL_GROUP
+    _MODEL_PARALLEL_GROUP = None
+    global _DATA_PARALLEL_GROUP
+    _DATA_PARALLEL_GROUP = None
