+++ NLP/GLM/mpu/grads.py
@@ -21,8 +21,8 @@
 import oneflow as flow
 # from oneflow._six import inf
 
-# from .initialize import get_model_parallel_group
-# from .initialize import get_model_parallel_rank
+# from .distribute import get_model_parallel_group
+# from .distribute import get_model_parallel_rank
 
 
 def clip_grad_norm(parameters, max_norm, norm_type=2):
