+++ NLP/GLM/mpu/mappings.py
@@ -15,7 +15,7 @@
 
 import torch
 
-from .initialize import get_model_parallel_group
+from .distribute import get_model_parallel_group
 from .utils import split_tensor_along_last_dim
