+++ NLP/GLM/mpu/random.py
@@ -46,10 +46,10 @@ def see_memory_usage(message, force=False):
         #input("Press Any Key To Continue ..")
 
 
-from .initialize import get_data_parallel_rank
-from .initialize import get_model_parallel_rank
-from .initialize import get_model_parallel_world_size
-from .initialize import get_model_parallel_group
+from .distribute import get_data_parallel_rank
+from .distribute import get_model_parallel_rank
+from .distribute import get_model_parallel_world_size
+from .distribute import get_model_parallel_group
 
 mp_rank = None #get_model_parallel_rank()
 mp_size = None #get_model_parallel_world_size()
