+++ NLP/GLM/mpu/transformer.py
@@ -23,7 +23,7 @@ import oneflow.nn.init as init
 from  oneflow.nn import LayerNorm
 from torch._C import device
 
-from .initialize import get_model_parallel_world_size
+from .distribute import get_model_parallel_world_size
 from .layers import ColumnParallelLinear
 from .layers import RowParallelLinear
 from .mappings import gather_from_model_parallel_region
